Generative AI/requirements.txt
langchain==0.0.184
langchain-core==0.1.30
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
huggingface-hub==0.14.1